+++ MoneyFi-backend/MoneyFi-budget/src/main/java/com/moneyfi/budget/repository/BudgetRepository.java
@@ -6,10 +6,10 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface BudgetRepository extends JpaRepository<BudgetModel, Integer> {
+public interface BudgetRepository extends JpaRepository<BudgetModel, Long> {
 
     @Query("select b from BudgetModel b where b.userId = :userId")
 //    @Query(nativeQuery = true, value = "exec getAllBudgetsByUserId @userId = :userId")
-    public List<BudgetModel> getBudgetsByUserId(int userId);
+    public List<BudgetModel> getBudgetsByUserId(Long userId);
 
 }
